FSUIPC Clang fixes

+++ src/blackmisc/simulation/fscommon/bcdconversions.cpp
@@ -31,7 +31,7 @@ namespace BlackMisc
             quint32 CBcdConversions::transponderCodeToBcd(const BlackMisc::Aviation::CTransponder &transponder)
             {
                 // FSX documentation is wrong, we need to use kHz + 2 digits, not Hz
-                quint32 t = transponder.getTransponderCode();
+                quint32 t = static_cast<quint32>(transponder.getTransponderCode());
                 t = dec2Bcd(t);
                 return t;
             }
